Refactor src/components/UserPortal/* from Jest to Vitest (#2982)
* chatroom tests migrated from jest to vitest

* chore: fixed undefined typescript _id error in chatroom component & add missing mutation in orginizationCard_

* chore: createDirectChat tests migrated from jest to vitest

* chore: creategroupchat tests migrated from jest to vitest

* chore: donationcard,orgcard, orgnavbar, orgsidebar tests migrated from jest to vitest

* chore: peoplecard, postcard, promotedPost, register, securedrouteforuser tests migrated from jest to vitest

* chore: startPostModal, usernavbar, userprofile, usersidebar, usersidebarorg tests migrated from jest to vitest
