+++ src/components/UserPortal/CreateDirectChat/CreateDirectChat.spec.tsx
@@ -1,36 +1,339 @@
 import React from 'react';
-
 import {
   act,
+  fireEvent,
   render,
   screen,
-  fireEvent,
   waitFor,
 } from '@testing-library/react';
-import { MockSubscriptionLink, MockedProvider } from '@apollo/react-testing';
+import { MockedProvider } from '@apollo/react-testing';
 import { I18nextProvider } from 'react-i18next';
+import { USERS_CONNECTION_LIST } from 'GraphQl/Queries/Queries';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { store } from 'state/store';
 import i18nForTest from 'utils/i18nForTest';
-import { CHATS_LIST, CHAT_BY_ID } from 'GraphQl/Queries/PlugInQueries';
+import Chat from '../../../screens/UserPortal/Chat/Chat';
 import {
+  CREATE_CHAT,
   MESSAGE_SENT_TO_CHAT,
-  SEND_MESSAGE_TO_CHAT,
 } from 'GraphQl/Mutations/OrganizationMutations';
-import ChatRoom from './ChatRoom';
-import { useLocalStorage } from 'utils/useLocalstorage';
-import { StaticMockLink } from 'utils/StaticMockLink';
-
+import { CHATS_LIST, CHAT_BY_ID } from 'GraphQl/Queries/PlugInQueries';
+import useLocalStorage from 'utils/useLocalstorage';
+import { vi } from 'vitest';
 const { setItem } = useLocalStorage();
 
-async function wait(ms = 100): Promise<void> {
-  await act(() => {
-    return new Promise((resolve) => {
-      setTimeout(resolve, ms);
-    });
-  });
-}
+/**
+ * Unit tests for the Create Direct Chat Modal functionality in the User Portal
+ *
+ * These tests cover the following scenarios:
+ * 1. Opening and closing the create new direct chat modal, ensuring proper UI elements
+ *    like dropdown, search input, and submit button are displayed and functional.
+ * 2. Creating a new direct chat, which includes testing the interaction with the add button,
+ *    submitting the chat, and closing the modal.
+ *
+ * Tests involve interacting with the modal's UI elements, performing actions like
+ * opening the dropdown, searching for users, clicking on the submit button, and closing
+ * the modal. GraphQL mocks are used for testing chat-related queries and mutations,
+ * ensuring that the modal behaves as expected when interacting with the GraphQL API.
+ */
+
+const UserConnectionListMock = [
+  {
+    request: {
+      query: USERS_CONNECTION_LIST,
+      variables: {
+        firstName_contains: '',
+        lastName_contains: '',
+      },
+    },
+    result: {
+      data: {
+        users: [
+          {
+            user: {
+              firstName: 'Deanne',
+              lastName: 'Marks',
+              image: null,
+              _id: '6589389d2caa9d8d69087487',
+              email: 'testuser8@example.com',
+              createdAt: '2023-04-13T04:53:17.742Z',
+              organizationsBlockedBy: [],
+              joinedOrganizations: [
+                {
+                  _id: '6537904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Queens',
+                    countryCode: 'US',
+                    dependentLocality: 'Some Dependent Locality',
+                    line1: '123 Coffee Street',
+                    line2: 'Apartment 501',
+                    postalCode: '11427',
+                    sortingCode: 'ABC-133',
+                    state: 'NYC',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+                {
+                  _id: '6637904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Staten Island',
+                    countryCode: 'US',
+                    dependentLocality: 'Some Dependent Locality',
+                    line1: '123 church Street',
+                    line2: 'Apartment 499',
+                    postalCode: '10301',
+                    sortingCode: 'ABC-122',
+                    state: 'NYC',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+                {
+                  _id: '6737904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Brooklyn',
+                    countryCode: 'US',
+                    dependentLocality: 'Sample Dependent Locality',
+                    line1: '123 Main Street',
+                    line2: 'Apt 456',
+                    postalCode: '10004',
+                    sortingCode: 'ABC-789',
+                    state: 'NY',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+                {
+                  _id: '6437904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Bronx',
+                    countryCode: 'US',
+                    dependentLocality: 'Some Dependent Locality',
+                    line1: '123 Random Street',
+                    line2: 'Apartment 456',
+                    postalCode: '10451',
+                    sortingCode: 'ABC-123',
+                    state: 'NYC',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+              ],
+              __typename: 'User',
+            },
+            appUserProfile: {
+              _id: '64378abd85308f171cf2993d',
+              adminFor: [],
+              isSuperAdmin: false,
+              createdOrganizations: [],
+              createdEvents: [],
+              eventAdmin: [],
+              __typename: 'AppUserProfile',
+            },
+            __typename: 'UserData',
+          },
+        ],
+      },
+    },
+  },
+  {
+    request: {
+      query: USERS_CONNECTION_LIST,
+      variables: {
+        firstName_contains: 'Disha',
+        lastName_contains: '',
+      },
+    },
+    result: {
+      data: {
+        users: [
+          {
+            user: {
+              firstName: 'Deanne',
+              lastName: 'Marks',
+              image: null,
+              _id: '6589389d2caa9d8d69087487',
+              email: 'testuser8@example.com',
+              createdAt: '2023-04-13T04:53:17.742Z',
+              organizationsBlockedBy: [],
+              joinedOrganizations: [
+                {
+                  _id: '6537904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Queens',
+                    countryCode: 'US',
+                    dependentLocality: 'Some Dependent Locality',
+                    line1: '123 Coffee Street',
+                    line2: 'Apartment 501',
+                    postalCode: '11427',
+                    sortingCode: 'ABC-133',
+                    state: 'NYC',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+                {
+                  _id: '6637904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Staten Island',
+                    countryCode: 'US',
+                    dependentLocality: 'Some Dependent Locality',
+                    line1: '123 church Street',
+                    line2: 'Apartment 499',
+                    postalCode: '10301',
+                    sortingCode: 'ABC-122',
+                    state: 'NYC',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+                {
+                  _id: '6737904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Brooklyn',
+                    countryCode: 'US',
+                    dependentLocality: 'Sample Dependent Locality',
+                    line1: '123 Main Street',
+                    line2: 'Apt 456',
+                    postalCode: '10004',
+                    sortingCode: 'ABC-789',
+                    state: 'NY',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+                {
+                  _id: '6437904485008f171cf29924',
+                  name: 'Unity Foundation',
+                  image: null,
+                  address: {
+                    city: 'Bronx',
+                    countryCode: 'US',
+                    dependentLocality: 'Some Dependent Locality',
+                    line1: '123 Random Street',
+                    line2: 'Apartment 456',
+                    postalCode: '10451',
+                    sortingCode: 'ABC-123',
+                    state: 'NYC',
+                    __typename: 'Address',
+                  },
+                  createdAt: '2023-04-13T05:16:52.827Z',
+                  creator: {
+                    _id: '64378abd85008f171cf2990d',
+                    firstName: 'Wilt',
+                    lastName: 'Shepherd',
+                    image: null,
+                    email: 'testsuperadmin@example.com',
+                    createdAt: '2023-04-13T04:53:17.742Z',
+                    __typename: 'User',
+                  },
+                  __typename: 'Organization',
+                },
+              ],
+              __typename: 'User',
+            },
+            appUserProfile: {
+              _id: '64378abd85308f171cf2993d',
+              adminFor: [],
+              isSuperAdmin: false,
+              createdOrganizations: [],
+              createdEvents: [],
+              eventAdmin: [],
+              __typename: 'AppUserProfile',
+            },
+            __typename: 'UserData',
+          },
+        ],
+      },
+    },
+  },
+];
 
 const MESSAGE_SENT_TO_CHAT_MOCK = [
   {
@@ -45,10 +348,8 @@ const MESSAGE_SENT_TO_CHAT_MOCK = [
         messageSentToChat: {
           _id: '668ec1f1364e03ac47a151',
           createdAt: '2024-07-10T17:16:33.248Z',
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
           messageContent: 'Test ',
+          type: 'STRING',
           replyTo: null,
           sender: {
             _id: '64378abd85008f171cf2990d',
@@ -74,10 +375,8 @@ const MESSAGE_SENT_TO_CHAT_MOCK = [
           _id: '668ec1f1df364e03ac47a151',
           createdAt: '2024-07-10T17:16:33.248Z',
           messageContent: 'Test ',
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
           replyTo: null,
+          type: 'STRING',
           sender: {
             _id: '64378abd85008f171cf2990d',
             firstName: 'Wilt',
@@ -102,10 +401,8 @@ const MESSAGE_SENT_TO_CHAT_MOCK = [
           _id: '668ec1f13603ac4697a151',
           createdAt: '2024-07-10T17:16:33.248Z',
           messageContent: 'Test ',
-          chatMessageBelongsTo: {
-            _id: '1',
-          },
           replyTo: null,
+          type: 'STRING',
           sender: {
             _id: '64378abd85008f171cf2990d',
             firstName: 'Wilt',
@@ -146,10 +443,11 @@ const CHAT_BY_ID_QUERY_MOCK = [
           name: '',
           messages: [
             {
-              _id: '4',
+              _id: '345678',
               createdAt: '345678908765',
               messageContent: 'Hello',
               replyTo: null,
+              type: 'STRING',
               sender: {
                 _id: '2',
                 firstName: 'Test',
@@ -205,10 +503,11 @@ const CHAT_BY_ID_QUERY_MOCK = [
           createdAt: '2345678903456',
           messages: [
             {
-              _id: '4',
+              _id: '345678',
               createdAt: '345678908765',
               messageContent: 'Hello',
               replyTo: null,
+              type: 'STRING',
               sender: {
                 _id: '2',
                 firstName: 'Test',
@@ -264,10 +563,11 @@ const CHAT_BY_ID_QUERY_MOCK = [
           createdAt: '2345678903456',
           messages: [
             {
-              _id: '4',
+              _id: '345678',
               createdAt: '345678908765',
               messageContent: 'Hello',
               replyTo: null,
+              type: 'STRING',
               sender: {
                 _id: '2',
                 firstName: 'Test',
@@ -334,6 +634,7 @@ const CHATS_LIST_MOCK = [
                 createdAt: '345678908765',
                 messageContent: 'Hello',
                 replyTo: null,
+                type: 'STRING',
                 sender: {
                   _id: '2',
                   firstName: 'Test',
@@ -405,6 +706,7 @@ const CHATS_LIST_MOCK = [
                 createdAt: '345678908765',
                 messageContent: 'Hello',
                 replyTo: null,
+                type: 'STRING',
                 sender: {
                   _id: '2',
                   firstName: 'Test',
@@ -490,6 +792,7 @@ const CHATS_LIST_MOCK = [
                 createdAt: '345678908765',
                 messageContent: 'Hello',
                 replyTo: null,
+                type: 'STRING',
                 sender: {
                   _id: '2',
                   firstName: 'Test',
@@ -561,6 +864,7 @@ const CHATS_LIST_MOCK = [
                 createdAt: '345678908765',
                 messageContent: 'Hello',
                 replyTo: null,
+                type: 'STRING',
                 sender: {
                   _id: '2',
                   firstName: 'Test',
@@ -646,6 +950,7 @@ const CHATS_LIST_MOCK = [
                 createdAt: '345678908765',
                 messageContent: 'Hello',
                 replyTo: null,
+                type: 'STRING',
                 sender: {
                   _id: '2',
                   firstName: 'Test',
@@ -717,6 +1022,165 @@ const CHATS_LIST_MOCK = [
                 createdAt: '345678908765',
                 messageContent: 'Hello',
                 replyTo: null,
+                type: 'STRING',
+                sender: {
+                  _id: '2',
+                  firstName: 'Test',
+                  lastName: 'User',
+                  email: 'test@example.com',
+                  image: '',
+                },
+              },
+            ],
+            users: [
+              {
+                _id: '1',
+                firstName: 'Disha',
+                lastName: 'Talreja',
+                email: 'disha@example.com',
+                image: '',
+              },
+              {
+                _id: '2',
+                firstName: 'Test',
+                lastName: 'User',
+                email: 'test@example.com',
+                image: '',
+              },
+              {
+                _id: '3',
+                firstName: 'Test',
+                lastName: 'User1',
+                email: 'test1@example.com',
+                image: '',
+              },
+              {
+                _id: '4',
+                firstName: 'Test',
+                lastName: 'User2',
+                email: 'test2@example.com',
+                image: '',
+              },
+              {
+                _id: '5',
+                firstName: 'Test',
+                lastName: 'User4',
+                email: 'test4@example.com',
+                image: '',
+              },
+            ],
+          },
+        ],
+      },
+    },
+  },
+  {
+    request: {
+      query: CHATS_LIST,
+      variables: {
+        id: '1',
+      },
+    },
+    result: {
+      data: {
+        chatsByUserId: [
+          {
+            _id: '65844efc814dhjmkdftyd4003db811c4',
+            isGroup: true,
+            creator: {
+              _id: '64378abd85008f171cf2990d',
+              firstName: 'Wilt',
+              lastName: 'Shepherd',
+              image: null,
+              email: 'testsuperadmin@example.com',
+              createdAt: '2023-04-13T04:53:17.742Z',
+              __typename: 'User',
+            },
+            organization: {
+              _id: 'pw3ertyuiophgfre45678',
+              name: 'rtyu',
+            },
+            createdAt: '2345678903456',
+            name: 'Test Group Chat',
+            messages: [
+              {
+                _id: '345678',
+                createdAt: '345678908765',
+                messageContent: 'Hello',
+                replyTo: null,
+                type: 'STRING',
+                sender: {
+                  _id: '2',
+                  firstName: 'Test',
+                  lastName: 'User',
+                  email: 'test@example.com',
+                  image: '',
+                },
+              },
+            ],
+            users: [
+              {
+                _id: '1',
+                firstName: 'Disha',
+                lastName: 'Talreja',
+                email: 'disha@example.com',
+                image: '',
+              },
+              {
+                _id: '2',
+                firstName: 'Test',
+                lastName: 'User',
+                email: 'test@example.com',
+                image: '',
+              },
+              {
+                _id: '3',
+                firstName: 'Test',
+                lastName: 'User1',
+                email: 'test1@example.com',
+                image: '',
+              },
+              {
+                _id: '4',
+                firstName: 'Test',
+                lastName: 'User2',
+                email: 'test2@example.com',
+                image: '',
+              },
+              {
+                _id: '5',
+                firstName: 'Test',
+                lastName: 'User4',
+                email: 'test4@example.com',
+                image: '',
+              },
+            ],
+          },
+          {
+            _id: '65844ewsedrffc814dd4003db811c4',
+            isGroup: true,
+            creator: {
+              _id: '64378abd85008f171cf2990d',
+              firstName: 'Wilt',
+              lastName: 'Shepherd',
+              image: null,
+              email: 'testsuperadmin@example.com',
+              createdAt: '2023-04-13T04:53:17.742Z',
+              __typename: 'User',
+            },
+            organization: {
+              _id: 'pw3ertyuiophgfre45678',
+              name: 'rtyu',
+            },
+            createdAt: '2345678903456',
+            name: 'Test Group Chat',
+            messages: [
+              {
+                _id: '345678',
+                createdAt: '345678908765',
+                messageContent: 'Hello',
+                replyTo: null,
+                type: 'STRING',
                 sender: {
                   _id: '2',
                   firstName: 'Test',
@@ -800,176 +1264,11 @@ const GROUP_CHAT_BY_ID_QUERY_MOCK = [
           name: 'Test Group Chat',
           messages: [
             {
-              _id: '2',
+              _id: '345678',
               createdAt: '345678908765',
               messageContent: 'Hello',
               replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '1',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              replyTo: null,
-              sender: {
-                _id: '2',
-                firstName: 'Test',
-                lastName: 'User',
-                email: 'test@example.com',
-                image: '',
-              },
-            },
-          ],
-          users: [
-            {
-              _id: '1',
-              firstName: 'Disha',
-              lastName: 'Talreja',
-              email: 'disha@example.com',
-              image: '',
-            },
-            {
-              _id: '2',
-              firstName: 'Test',
-              lastName: 'User',
-              email: 'test@example.com',
-              image: '',
-            },
-            {
-              _id: '3',
-              firstName: 'Test',
-              lastName: 'User1',
-              email: 'test1@example.com',
-              image: '',
-            },
-            {
-              _id: '4',
-              firstName: 'Test',
-              lastName: 'User2',
-              email: 'test2@example.com',
-              image: '',
-            },
-            {
-              _id: '5',
-              firstName: 'Test',
-              lastName: 'User4',
-              email: 'test4@example.com',
-              image: '',
-            },
-          ],
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: CHAT_BY_ID,
-      variables: {
-        id: '1',
-      },
-    },
-    result: {
-      data: {
-        chatById: {
-          _id: '65844efc814dd4003db811c4',
-          isGroup: true,
-          creator: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: null,
-            email: 'testsuperadmin@example.com',
-            createdAt: '2023-04-13T04:53:17.742Z',
-            __typename: 'User',
-          },
-          organization: {
-            _id: 'pw3ertyuiophgfre45678',
-            name: 'rtyu',
-          },
-          createdAt: '2345678903456',
-          name: 'Test Group Chat',
-          messages: [
-            {
-              _id: '1',
-              createdAt: '345678908765',
-              messageContent: 'Hello',
-              replyTo: null,
+              type: 'STRING',
               sender: {
                 _id: '2',
                 firstName: 'Test',
@@ -1022,565 +1321,191 @@ const GROUP_CHAT_BY_ID_QUERY_MOCK = [
   },
 ];
 
-const SEND_MESSAGE_TO_CHAT_MOCK = [
+const CREATE_CHAT_MUTATION_MOCK = [
   {
     request: {
-      query: SEND_MESSAGE_TO_CHAT,
+      query: CREATE_CHAT,
       variables: {
-        chatId: '1',
-        replyTo: '4',
-        messageContent: 'Test reply message',
+        organizationId: undefined,
+        userIds: ['1', '6589389d2caa9d8d69087487'],
+        isGroup: false,
       },
     },
     result: {
       data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          replyTo: null,
-          sender: {
+        createChat: {
+          _id: '1',
+          createdAt: '2345678903456',
+          isGroup: false,
+          creator: {
             _id: '64378abd85008f171cf2990d',
             firstName: 'Wilt',
             lastName: 'Shepherd',
-            image: '',
+            image: null,
+            email: 'testsuperadmin@example.com',
+            createdAt: '2023-04-13T04:53:17.742Z',
+            __typename: 'User',
           },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '4',
-        messageContent: 'Test reply message',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '1',
-        messageContent: 'Test reply message',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: undefined,
-        messageContent: 'Hello',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: '345678',
-        messageContent: 'Test reply message',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: SEND_MESSAGE_TO_CHAT,
-      variables: {
-        chatId: '1',
-        replyTo: undefined,
-        messageContent: 'Test message',
-      },
-    },
-    result: {
-      data: {
-        sendMessageToChat: {
-          _id: '668ec1f1364e03ac47a151',
-          createdAt: '2024-07-10T17:16:33.248Z',
-          messageContent: 'Test ',
-          replyTo: null,
-          sender: {
-            _id: '64378abd85008f171cf2990d',
-            firstName: 'Wilt',
-            lastName: 'Shepherd',
-            image: '',
-          },
-          updatedAt: '2024-07-10',
+          organization: null,
+          name: '',
+          messages: [
+            {
+              _id: '345678',
+              createdAt: '345678908765',
+              messageContent: 'Hello',
+              replyTo: null,
+              type: 'STRING',
+              sender: {
+                _id: '2',
+                firstName: 'Test',
+                lastName: 'User',
+                email: 'test@example.com',
+                image: '',
+              },
+            },
+          ],
+          users: [
+            {
+              _id: '1',
+              firstName: 'Disha',
+              lastName: 'Talreja',
+              email: 'disha@example.com',
+              image: '',
+            },
+            {
+              _id: '2',
+              firstName: 'Test',
+              lastName: 'User',
+              email: 'test@example.com',
+              image: '',
+            },
+          ],
         },
       },
     },
   },
 ];
 
-describe('Testing Chatroom Component [User Portal]', () => {
-  window.HTMLElement.prototype.scrollIntoView = jest.fn();
+async function wait(ms = 100): Promise<void> {
+  await act(() => {
+    return new Promise((resolve) => {
+      setTimeout(resolve, ms);
+    });
+  });
+}
 
-  test('Chat room should display fallback content if no chat is active', async () => {
-    const mocks = [
-      ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
-      ...CHATS_LIST_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
-    ];
-    render(
-      <MockedProvider addTypename={false} mocks={mocks}>
-        <BrowserRouter>
-          <Provider store={store}>
-            <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="" />
-            </I18nextProvider>
-          </Provider>
-        </BrowserRouter>
-      </MockedProvider>,
-    );
-    await wait();
-    expect(await screen.findByTestId('noChatSelected')).toBeInTheDocument();
+describe('Testing Create Direct Chat Modal [User Portal]', () => {
+  window.HTMLElement.prototype.scrollIntoView = vi.fn();
+
+  Object.defineProperty(window, 'matchMedia', {
+    writable: true,
+    value: vi.fn().mockImplementation((query) => ({
+      matches: false,
+      media: query,
+      onchange: null,
+      addListener: vi.fn(), // Deprecated
+      removeListener: vi.fn(), // Deprecated
+      addEventListener: vi.fn(),
+      removeEventListener: vi.fn(),
+      dispatchEvent: vi.fn(),
+    })),
   });
 
-  test('Selected contact is direct chat', async () => {
-    const link = new MockSubscriptionLink();
-    const mocks = [
-      ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
-      ...CHATS_LIST_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
-    ];
-    render(
-      <MockedProvider addTypename={false} mocks={mocks} link={link}>
-        <BrowserRouter>
-          <Provider store={store}>
-            <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="1" />
-            </I18nextProvider>
-          </Provider>
-        </BrowserRouter>
-      </MockedProvider>,
-    );
-    await wait();
-  });
-
-  test('send message direct chat', async () => {
-    setItem('userId', '2');
-    const mocks = [
-      ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
-      ...CHATS_LIST_MOCK,
-      ...GROUP_CHAT_BY_ID_QUERY_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
-    ];
-    const link2 = new StaticMockLink(mocks, true);
-    render(
-      <MockedProvider addTypename={false} link={link2}>
-        <BrowserRouter>
-          <Provider store={store}>
-            <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="1" />
-            </I18nextProvider>
-          </Provider>
-        </BrowserRouter>
-      </MockedProvider>,
-    );
-    await wait();
-
-    const input = (await screen.findByTestId(
-      'messageInput',
-    )) as HTMLInputElement;
-
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Hello' } });
-    });
-    expect(input.value).toBe('Hello');
-
-    const sendBtn = await screen.findByTestId('sendMessage');
-
-    expect(sendBtn).toBeInTheDocument();
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-    await waitFor(() => {
-      expect(input.value).toBeFalsy();
-    });
-
-    const messages = await screen.findAllByTestId('message');
-
-    console.log('MESSAGES', messages);
-
-    expect(messages.length).not.toBe(0);
-
-    act(() => {
-      fireEvent.mouseOver(messages[0]);
-    });
-
-    await waitFor(async () => {
-      expect(await screen.findByTestId('moreOptions')).toBeInTheDocument();
-    });
-
-    const moreOptionsBtn = await screen.findByTestId('dropdown');
-    act(() => {
-      fireEvent.click(moreOptionsBtn);
-    });
-
-    const replyBtn = await screen.findByTestId('replyBtn');
-
-    act(() => {
-      fireEvent.click(replyBtn);
-    });
-
-    const replyMsg = await screen.findByTestId('replyMsg');
-
-    await waitFor(() => {
-      expect(replyMsg).toBeInTheDocument();
-    });
-
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Test reply message' } });
-    });
-    expect(input.value).toBe('Test reply message');
-
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-
-    await wait(400);
-  });
-
-  test('send message direct chat when userId is different', async () => {
-    setItem('userId', '8');
-    const mocks = [
+  it('Open and close create new direct chat modal', async () => {
+    const mock = [
       ...GROUP_CHAT_BY_ID_QUERY_MOCK,
       ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
+      ...UserConnectionListMock,
       ...CHATS_LIST_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
+      ...CHAT_BY_ID_QUERY_MOCK,
+      ...CREATE_CHAT_MUTATION_MOCK,
     ];
-    const link2 = new StaticMockLink(mocks, true);
     render(
-      <MockedProvider addTypename={false} link={link2}>
+      <MockedProvider addTypename={false} mocks={mock}>
         <BrowserRouter>
           <Provider store={store}>
             <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="1" />
+              <Chat />
             </I18nextProvider>
           </Provider>
         </BrowserRouter>
       </MockedProvider>,
     );
+
     await wait();
 
-    const input = (await screen.findByTestId(
-      'messageInput',
+    const dropdown = await screen.findByTestId('dropdown');
+    expect(dropdown).toBeInTheDocument();
+    fireEvent.click(dropdown);
+    const newDirectChatBtn = await screen.findByTestId('newDirectChat');
+    expect(newDirectChatBtn).toBeInTheDocument();
+    fireEvent.click(newDirectChatBtn);
+
+    const submitBtn = await screen.findByTestId('submitBtn');
+    expect(submitBtn).toBeInTheDocument();
+
+    const searchInput = (await screen.findByTestId(
+      'searchUser',
     )) as HTMLInputElement;
+    expect(searchInput).toBeInTheDocument();
 
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Hello' } });
-    });
-    expect(input.value).toBe('Hello');
+    fireEvent.change(searchInput, { target: { value: 'Disha' } });
 
-    const sendBtn = await screen.findByTestId('sendMessage');
+    expect(searchInput.value).toBe('Disha');
 
-    expect(sendBtn).toBeInTheDocument();
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-    await waitFor(() => {
-      expect(input.value).toBeFalsy();
-    });
+    fireEvent.click(submitBtn);
 
-    const messages = await screen.findAllByTestId('message');
+    const closeButton = screen.getByRole('button', { name: /close/i });
+    expect(closeButton).toBeInTheDocument();
 
-    console.log('MESSAGES', messages);
-
-    expect(messages.length).not.toBe(0);
-
-    act(() => {
-      fireEvent.mouseOver(messages[0]);
-    });
-
-    await waitFor(async () => {
-      expect(await screen.findByTestId('moreOptions')).toBeInTheDocument();
-    });
-
-    const moreOptionsBtn = await screen.findByTestId('dropdown');
-    act(() => {
-      fireEvent.click(moreOptionsBtn);
-    });
-
-    const replyBtn = await screen.findByTestId('replyBtn');
-
-    act(() => {
-      fireEvent.click(replyBtn);
-    });
-
-    const replyMsg = await screen.findByTestId('replyMsg');
-
-    await waitFor(() => {
-      expect(replyMsg).toBeInTheDocument();
-    });
-
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Test reply message' } });
-    });
-    expect(input.value).toBe('Test reply message');
-
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-
-    await wait(400);
+    fireEvent.click(closeButton);
   });
 
-  test('Selected contact is group chat', async () => {
-    const mocks = [
+  it('create new direct chat', async () => {
+    setItem('userId', '1');
+    const mock = [
+      ...GROUP_CHAT_BY_ID_QUERY_MOCK,
       ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
+      ...UserConnectionListMock,
       ...CHATS_LIST_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
+      ...CHAT_BY_ID_QUERY_MOCK,
+      ...CREATE_CHAT_MUTATION_MOCK,
     ];
     render(
-      <MockedProvider addTypename={false} mocks={mocks}>
+      <MockedProvider addTypename={false} mocks={mock}>
         <BrowserRouter>
           <Provider store={store}>
             <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="1" />
+              <Chat />
             </I18nextProvider>
           </Provider>
         </BrowserRouter>
       </MockedProvider>,
     );
-    await wait();
-  });
 
-  test('send message group chat', async () => {
-    const mocks = [
-      ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
-      ...CHATS_LIST_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
-    ];
-    render(
-      <MockedProvider addTypename={false} mocks={mocks}>
-        <BrowserRouter>
-          <Provider store={store}>
-            <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="1" />
-            </I18nextProvider>
-          </Provider>
-        </BrowserRouter>
-      </MockedProvider>,
-    );
     await wait();
 
-    const input = (await screen.findByTestId(
-      'messageInput',
-    )) as HTMLInputElement;
+    const dropdown = await screen.findByTestId('dropdown');
+    expect(dropdown).toBeInTheDocument();
+    fireEvent.click(dropdown);
+    const newDirectChatBtn = await screen.findByTestId('newDirectChat');
+    expect(newDirectChatBtn).toBeInTheDocument();
+    fireEvent.click(newDirectChatBtn);
 
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Test message' } });
-    });
-    expect(input.value).toBe('Test message');
-
-    const sendBtn = await screen.findByTestId('sendMessage');
-
-    expect(sendBtn).toBeInTheDocument();
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-    await waitFor(() => {
-      expect(input.value).toBeFalsy();
+    const addBtn = await screen.findAllByTestId('addBtn');
+    waitFor(() => {
+      expect(addBtn[0]).toBeInTheDocument();
     });
 
-    const messages = await screen.findAllByTestId('message');
+    fireEvent.click(addBtn[0]);
 
-    expect(messages.length).not.toBe(0);
+    const closeButton = screen.getByRole('button', { name: /close/i });
+    expect(closeButton).toBeInTheDocument();
 
-    act(() => {
-      fireEvent.mouseOver(messages[0]);
-    });
+    fireEvent.click(closeButton);
 
-    expect(await screen.findByTestId('moreOptions')).toBeInTheDocument();
+    await new Promise(process.nextTick);
 
-    const moreOptionsBtn = await screen.findByTestId('dropdown');
-    act(() => {
-      fireEvent.click(moreOptionsBtn);
-    });
-
-    const replyBtn = await screen.findByTestId('replyBtn');
-
-    act(() => {
-      fireEvent.click(replyBtn);
-    });
-
-    const replyMsg = await screen.findByTestId('replyMsg');
-
-    await waitFor(() => {
-      expect(replyMsg).toBeInTheDocument();
-    });
-
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Test reply message' } });
-    });
-    expect(input.value).toBe('Test reply message');
-
-    const closeReplyBtn = await screen.findByTestId('closeReply');
-
-    expect(closeReplyBtn).toBeInTheDocument();
-
-    fireEvent.click(closeReplyBtn);
-
-    await wait(500);
-  });
-
-  test('reply to message', async () => {
-    const mocks = [
-      ...MESSAGE_SENT_TO_CHAT_MOCK,
-      ...CHAT_BY_ID_QUERY_MOCK,
-      ...CHATS_LIST_MOCK,
-      ...SEND_MESSAGE_TO_CHAT_MOCK,
-    ];
-    const link2 = new StaticMockLink(mocks, true);
-    render(
-      <MockedProvider addTypename={false} link={link2}>
-        <BrowserRouter>
-          <Provider store={store}>
-            <I18nextProvider i18n={i18nForTest}>
-              <ChatRoom selectedContact="1" />
-            </I18nextProvider>
-          </Provider>
-        </BrowserRouter>
-      </MockedProvider>,
-    );
     await wait();
-
-    const input = (await screen.findByTestId(
-      'messageInput',
-    )) as HTMLInputElement;
-
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Test message' } });
-    });
-    expect(input.value).toBe('Test message');
-
-    const sendBtn = await screen.findByTestId('sendMessage');
-
-    expect(sendBtn).toBeInTheDocument();
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-    await waitFor(() => {
-      expect(input.value).toBeFalsy();
-    });
-
-    const messages = await screen.findAllByTestId('message');
-
-    expect(messages.length).not.toBe(0);
-
-    act(() => {
-      fireEvent.mouseOver(messages[0]);
-    });
-
-    expect(await screen.findByTestId('moreOptions')).toBeInTheDocument();
-
-    const moreOptionsBtn = await screen.findByTestId('dropdown');
-    act(() => {
-      fireEvent.click(moreOptionsBtn);
-    });
-
-    const replyBtn = await screen.findByTestId('replyBtn');
-
-    act(() => {
-      fireEvent.click(replyBtn);
-    });
-
-    const replyMsg = await screen.findByTestId('replyMsg');
-
-    await waitFor(() => {
-      expect(replyMsg).toBeInTheDocument();
-    });
-
-    act(() => {
-      fireEvent.change(input, { target: { value: 'Test reply message' } });
-    });
-    expect(input.value).toBe('Test reply message');
-
-    act(() => {
-      fireEvent.click(sendBtn);
-    });
-
-    await wait(400);
   });
 });
