+++ src/components/UserPortal/CreateGroupChat/CreateGroupChat.spec.tsx
@@ -26,6 +26,24 @@ import {
 import { CHATS_LIST, CHAT_BY_ID } from 'GraphQl/Queries/PlugInQueries';
 import useLocalStorage from 'utils/useLocalstorage';
 import userEvent from '@testing-library/user-event';
+import { vi } from 'vitest';
+
+/**
+ * Unit tests for the Create Group Chat Modal functionality in the User Portal
+ *
+ * These tests cover the following scenarios:
+ * 1. Opening and closing the create new group chat modal, ensuring proper UI elements
+ *    like the dropdown, new group chat button, and close button are displayed and functional.
+ * 2. Creating a new group chat by interacting with the group name input field, organization
+ *    selection, and submission process. It also ensures that the create button is properly
+ *    triggered after filling out the required fields.
+ * 3. Adding and removing users from the group chat, testing the interactions with the add
+ *    and remove buttons, and verifying the submit button and search functionality for user selection.
+ *
+ * GraphQL mocks are used to simulate chat-related queries and mutations. The tests ensure that
+ * the modal behaves correctly in various user interaction scenarios, including handling of form
+ * fields, user management, and modal navigation.
+ */
 
 const { setItem } = useLocalStorage();
 
@@ -2212,23 +2230,23 @@ async function wait(ms = 100): Promise<void> {
 }
 
 describe('Testing Create Group Chat Modal [User Portal]', () => {
-  window.HTMLElement.prototype.scrollIntoView = jest.fn();
+  window.HTMLElement.prototype.scrollIntoView = vi.fn();
 
   Object.defineProperty(window, 'matchMedia', {
     writable: true,
-    value: jest.fn().mockImplementation((query) => ({
+    value: vi.fn().mockImplementation((query) => ({
       matches: false,
       media: query,
       onchange: null,
-      addListener: jest.fn(), // Deprecated
-      removeListener: jest.fn(), // Deprecated
-      addEventListener: jest.fn(),
-      removeEventListener: jest.fn(),
-      dispatchEvent: jest.fn(),
+      addListener: vi.fn(), // Deprecated
+      removeListener: vi.fn(), // Deprecated
+      addEventListener: vi.fn(),
+      removeEventListener: vi.fn(),
+      dispatchEvent: vi.fn(),
     })),
   });
 
-  test('open and close create new direct chat modal', async () => {
+  it('open and close create new direct chat modal', async () => {
     const mock = [
       ...USER_JOINED_ORG_MOCK,
       ...GROUP_CHAT_BY_ID_QUERY_MOCK,
@@ -2266,7 +2284,7 @@ describe('Testing Create Group Chat Modal [User Portal]', () => {
     fireEvent.click(closeButton);
   });
 
-  test('create new group chat', async () => {
+  it('create new group chat', async () => {
     const mock = [
       ...USER_JOINED_ORG_MOCK,
       ...GROUP_CHAT_BY_ID_QUERY_MOCK,
@@ -2366,7 +2384,7 @@ describe('Testing Create Group Chat Modal [User Portal]', () => {
     // });
   }, 3000);
 
-  test('add and remove user', async () => {
+  it('add and remove user', async () => {
     setItem('userId', '1');
     const mock = [
       ...USER_JOINED_ORG_MOCK,
