+++ src/components/UserPortal/DonationCard/DonationCard.spec.tsx
@@ -11,6 +11,19 @@ import { StaticMockLink } from 'utils/StaticMockLink';
 import DonationCard from './DonationCard';
 import { type InterfaceDonationCardProps } from 'screens/UserPortal/Donate/Donate';
 
+/**
+ * Unit test for the DonationCard component in the User Portal
+ *
+ * This test ensures that the DonationCard component renders correctly when provided with
+ * the required props. It uses the MockedProvider to mock any GraphQL queries and the
+ * StaticMockLink to simulate the network layer. The component is wrapped with necessary
+ * providers such as BrowserRouter, Redux store, and i18n provider to simulate the environment
+ * in which the component operates.
+ *
+ * The test specifically checks if the component renders without errors, though more tests
+ * can be added in the future to validate interactions and state changes based on user actions.
+ */
+
 const link = new StaticMockLink([], true);
 
 async function wait(ms = 100): Promise<void> {
@@ -31,7 +44,7 @@ const props: InterfaceDonationCardProps = {
 };
 
 describe('Testing ContactCard Component [User Portal]', () => {
-  test('Component should be rendered properly', async () => {
+  it('Component should be rendered properly', async () => {
     render(
       <MockedProvider addTypename={false} link={link}>
         <BrowserRouter>
