+++ src/components/UserPortal/EventCard/EventCard.spec.tsx
@@ -11,7 +11,6 @@ import { Provider } from 'react-redux';
 import { store } from 'state/store';
 import { StaticMockLink } from 'utils/StaticMockLink';
 import userEvent from '@testing-library/user-event';
-import { debug } from 'jest-preview';
 import useLocalStorage from 'utils/useLocalstorage';
 
 const { setItem } = useLocalStorage();
@@ -66,7 +65,7 @@ describe('Testing Event Card In User portal', () => {
     ],
   };
 
-  test('The card should be rendered properly, and all the details should be displayed correct', async () => {
+  it('The card should be rendered properly, and all the details should be displayed correct', async () => {
     const { queryByText } = render(
       <MockedProvider addTypename={false} link={link}>
         <BrowserRouter>
@@ -79,7 +78,6 @@ describe('Testing Event Card In User portal', () => {
         </BrowserRouter>
       </MockedProvider>,
     );
-    debug();
     await waitFor(() => expect(queryByText('Test Event')).toBeInTheDocument());
     await waitFor(() =>
       expect(queryByText('This is a test event')).toBeInTheDocument(),
@@ -105,7 +103,7 @@ describe('Testing Event Card In User portal', () => {
     await waitFor(() => expect(queryByText('Register')).toBeInTheDocument());
   });
 
-  test('When the user is already registered', async () => {
+  it('When the user is already registered', async () => {
     setItem('userId', '234');
     const { queryByText } = render(
       <MockedProvider addTypename={false} link={link}>
@@ -124,7 +122,7 @@ describe('Testing Event Card In User portal', () => {
     );
   });
 
-  test('Handle register should work properly', async () => {
+  it('Handle register should work properly', async () => {
     setItem('userId', '456');
     const { queryByText } = render(
       <MockedProvider addTypename={false} link={link}>
@@ -173,7 +171,7 @@ describe('Event card when start and end time are not given', () => {
     ],
   };
 
-  test('Card is rendered correctly', async () => {
+  it('Card is rendered correctly', async () => {
     const { container } = render(
       <MockedProvider addTypename={false} link={link}>
         <BrowserRouter>
