+++ src/components/UserPortal/OrganizationCard/OrganizationCard.spec.tsx
@@ -18,15 +18,33 @@ import useLocalStorage from 'utils/useLocalstorage';
 import {
   SEND_MEMBERSHIP_REQUEST,
   JOIN_PUBLIC_ORGANIZATION,
+  CANCEL_MEMBERSHIP_REQUEST,
 } from 'GraphQl/Mutations/OrganizationMutations';
 import { toast } from 'react-toastify';
+import { vi } from 'vitest';
 
+/**
+ * Unit tests for the OrganizationCard component in the User Portal
+ *
+ * These tests validate the behavior and rendering of the OrganizationCard component.
+ * The tests ensure the component displays properly with various states and that interactions
+ * such as sending membership requests and visiting organizations work as expected.
+ *
+ * 1. **Component should be rendered properly**: Tests if the component renders correctly with the provided props.
+ * 2. **Component should render properly with an image**: Verifies the component's behavior when an organization image is available.
+ * 3. **Visit organization**: Simulates a click on the "manage" button and verifies that the user is redirected to the correct organization page.
+ * 4. **Send membership request**: Tests if the membership request is successfully sent and verifies the success toast message.
+ * 5. **Send membership request to a public organization**: Validates sending a membership request to a public organization and verifies multiple success toast messages.
+ * 6. **Withdraw membership request**: Simulates withdrawing a membership request and verifies that the button works as expected.
+ *
+ * Mocked GraphQL queries and mutations are used to simulate the backend behavior for testing.
+ */
 const { getItem } = useLocalStorage();
 
-jest.mock('react-toastify', () => ({
+vi.mock('react-toastify', () => ({
   toast: {
-    success: jest.fn(),
-    error: jest.fn(),
+    success: vi.fn(),
+    error: vi.fn(),
   },
 }));
 
@@ -146,6 +164,21 @@ const MOCKS = [
       },
     },
   },
+  {
+    request: {
+      query: CANCEL_MEMBERSHIP_REQUEST,
+      variables: {
+        membershipRequestId: '56gheqyr7deyfuiwfewifruy8',
+      },
+    },
+    result: {
+      data: {
+        cancelMembershipRequest: {
+          _id: '56gheqyr7deyfuiwfewifruy8',
+        },
+      },
+    },
+  },
 ];
 
 const link = new StaticMockLink(MOCKS, true);
@@ -189,7 +222,7 @@ let props = {
 };
 
 describe('Testing OrganizationCard Component [User Portal]', () => {
-  test('Component should be rendered properly', async () => {
+  it('Component should be rendered properly', async () => {
     render(
       <MockedProvider addTypename={false} link={link}>
         <BrowserRouter>
@@ -205,7 +238,7 @@ describe('Testing OrganizationCard Component [User Portal]', () => {
     await wait();
   });
 
-  test('Component should be rendered properly if organization Image is not undefined', async () => {
+  it('Component should be rendered properly if organization Image is not undefined', async () => {
     props = {
       ...props,
       image: 'organizationImage',
@@ -226,7 +259,7 @@ describe('Testing OrganizationCard Component [User Portal]', () => {
     await wait();
   });
 
-  test('Visit organization', async () => {
+  it('Visit organization', async () => {
     const cardProps = {
       ...props,
       id: '3',
@@ -258,7 +291,7 @@ describe('Testing OrganizationCard Component [User Portal]', () => {
     expect(window.location.pathname).toBe(`/user/organization/${cardProps.id}`);
   });
 
-  test('Send membership request', async () => {
+  it('Send membership request', async () => {
     props = {
       ...props,
       image: 'organizationImage',
@@ -286,7 +319,7 @@ describe('Testing OrganizationCard Component [User Portal]', () => {
     expect(toast.success).toHaveBeenCalledWith('MembershipRequestSent');
   });
 
-  test('send membership request to public org', async () => {
+  it('send membership request to public org', async () => {
     const cardProps = {
       ...props,
       id: '2',
@@ -316,13 +349,21 @@ describe('Testing OrganizationCard Component [User Portal]', () => {
     expect(toast.success).toHaveBeenCalledTimes(2);
   });
 
-  test('withdraw membership request', async () => {
+  it('withdraw membership request', async () => {
     const cardProps = {
       ...props,
       id: '3',
       image: 'organizationImage',
       userRegistrationRequired: true,
       membershipRequestStatus: 'pending',
+      membershipRequests: [
+        {
+          _id: '56gheqyr7deyfuiwfewifruy8',
+          user: {
+            _id: getItem('userId'),
+          },
+        },
+      ],
     };
 
     render(
