+++ src/components/UserPortal/SecuredRouteForUser/SecuredRouteForUser.spec.tsx
@@ -4,10 +4,20 @@ import { render, screen, waitFor } from '@testing-library/react';
 import SecuredRouteForUser from './SecuredRouteForUser';
 import useLocalStorage from 'utils/useLocalstorage';
 
+/**
+ * Unit tests for SecuredRouteForUser component:
+ *
+ * 1. **Logged-in user**: Verifies that the route renders when 'IsLoggedIn' is set to 'TRUE'.
+ * 2. **Not logged-in user**: Ensures redirection to the login page when 'IsLoggedIn' is 'FALSE'.
+ * 3. **Logged-in user with admin access**: Checks that the route renders for a logged-in user with 'AdminFor' set (i.e., admin of an organization).
+ *
+ * LocalStorage values like 'IsLoggedIn' and 'AdminFor' are set to simulate different user states.
+ */
+
 const { setItem } = useLocalStorage();
 
 describe('SecuredRouteForUser', () => {
-  test('renders the route when the user is logged in', () => {
+  it('renders the route when the user is logged in', () => {
     // Set the 'IsLoggedIn' value to 'TRUE' in localStorage to simulate a logged-in user and do not set 'AdminFor' so that it remains undefined.
     setItem('IsLoggedIn', 'TRUE');
 
@@ -31,7 +41,7 @@ describe('SecuredRouteForUser', () => {
     expect(screen.getByTestId('organizations-content')).toBeInTheDocument();
   });
 
-  test('redirects to /user when the user is not logged in', async () => {
+  it('redirects to /user when the user is not logged in', async () => {
     // Set the user as not logged in in local storage
     setItem('IsLoggedIn', 'FALSE');
 
@@ -58,7 +68,7 @@ describe('SecuredRouteForUser', () => {
     });
   });
 
-  test('renders the route when the user is logged in and user is ADMIN', () => {
+  it('renders the route when the user is logged in and user is ADMIN', () => {
     // Set the 'IsLoggedIn' value to 'TRUE' in localStorage to simulate a logged-in user and set 'AdminFor' to simulate ADMIN of some Organization.
     setItem('IsLoggedIn', 'TRUE');
     setItem('AdminFor', [
